Fix BarrierOperator hidden overload (#1)

+++ cpp/src/command_classes/BarrierOperator.cpp
@@ -135,7 +135,7 @@ bool BarrierOperator::RequestState
 bool BarrierOperator::RequestValue
 (
 		uint32 const _requestFlags,
-		uint16 const _index,
+		uint8 const _index,
 		uint8 const _instance,
 		Driver::MsgQueue const _queue
 )
